Reorganize deprecated tests
