+++ test_crates/type_marked_deprecated/new/src/lib.rs
@@ -1,179 +1,3 @@
-// STRUCTS
-
-// These structs did not have the #[deprecated] attribute in the old version.
-// Addition of the attribute should be reported by this rule.
-
-#[deprecated]
-pub struct StructToDeprecatedStruct {
-    bar: u64,
-}
-
-#[deprecated = "This attribute was added"]
-pub struct StructToDeprecatedMessageStruct {
-    bar: u64,
-}
-
-// These structs had the #[deprecated] attribute in the old version. Changes of
-// the attribute, including deletion, should NOT be reported by this rule.
-
-pub struct DeprecatedStructToStruct {
-    bar: u64,
-}
-
-#[deprecated = "This message was added"]
-pub struct DeprecatedStructToDeprecatedMessageStruct {
-    bar: u64,
-}
-
-// These structs had the #[deprecated] attribute in the old version.
-// They also included the user-defined warning message. Changes of
-// the attribute, including deletion, should NOT be reported by this rule.
-
-pub struct DeprecatedMessageStructToStruct {
-    bar: u64,
-}
-
-#[deprecated]
-pub struct DeprecatedMessageStructToDeprecatedStruct {
-    bar: u64,
-}
-
-#[deprecated = "This message was changed"]
-pub struct DeprecatedMessageStructToDeprecatedMessageStruct {
-    bar: u64,
-}
-
-// This struct is private and should NOT be reported by this rule.
-
-#[deprecated]
-struct DeprecatedPrivateStruct {
-    bar: u64,
-}
-
-// This struct was added in the new version of the crate with its attribute.
-// It should NOT be reported by this rule to avoid duplicate lints.
-// It should be reported as a new pub type that is part of the crate's API.
-// This might seem like a nonsensical test but is a valid edge case.
-
-#[deprecated]
-pub struct DeprecatedNewStruct {
-    bar: u64,
-}
-
-// ENUMS
-
-// These enums did not have the #[deprecated] attribute in the old version.
-// Addition of the attribute should be reported.
-
-#[deprecated]
-pub enum EnumToDeprecatedEnum {
-    First,
-}
-
-#[deprecated = "This attribute was added"]
-pub enum EnumToDeprecatedMessageEnum {
-    First,
-}
-
-// These structs had the #[deprecated] attribute in the old version.
-// They also included the user-defined warning message. Changes of
-// the attribute, including deletion, should NOT be reported by this rule.
-
-pub enum DeprecatedEnumToEnum {
-    First,
-}
-
-#[deprecated = "This message was added"]
-pub enum DeprecatedEnumToDeprecatedMessageEnum {
-    First,
-}
-
-pub enum DeprecatedMessageEnumToEnum {
-    First,
-}
-
-#[deprecated]
-pub enum DeprecatedMessageEnumToDeprecatedEnum {
-    First,
-}
-
-#[deprecated = "This message was changed"]
-pub enum DeprecatedMessageEnumToDeprecatedMessageEnum {
-    First,
-}
-
-// This enum is private and should NOT be reported.
-
-#[deprecated]
-enum DeprecatedPrivateEnum {
-    First,
-}
-
-// This enum was added in the new version of the crate with its attribute.
-// It should NOT be reported by this rule to avoid duplicate lints.
-// It should be reported as a new pub type that is part of the crate's API.
-// This might seem like a nonsensical test but is a valid edge case.
-
-#[deprecated]
-pub enum DeprecatedNewEnum {
-    First,
-}
-
-// UNION
-
-// These unions did not have the #[deprecated] attribute in the old version.
-// Addition of the attribute should be reported.
-
-#[deprecated]
-pub union UnionToDeprecatedUnion {
-    bar: u64,
-}
-
-#[deprecated = "This attribute was added"]
-pub union UnionToDeprecatedMessageUnion {
-    bar: u64,
-}
-
-// These structs had the #[deprecated] attribute in the old version.
-// They also included the user-defined warning message. Changes of
-// the attribute, including deletion, should NOT be reported by this rule.
-
-pub union DeprecatedUnionToUnion {
-    bar: u64,
-}
-
-#[deprecated = "This message was added"]
-pub union DeprecatedUnionToDeprecatedMessageUnion {
-    bar: u64,
-}
-
-pub union DeprecatedMessageUnionToUnion {
-    bar: u64,
-}
-
-#[deprecated]
-pub union DeprecatedMessageUnionToDeprecatedUnion {
-    bar: u64,
-}
-
-#[deprecated = "This message was changed"]
-pub union DeprecatedMessageUnionToDeprecatedMessageUnion {
-    bar: u64,
-}
-
-// This union is private and should NOT be reported.
-
-#[deprecated]
-union DeprecatedPrivateUnion {
-    bar: u64,
-}
-
-// This union was added in the new version of the crate with its attribute.
-// It should NOT be reported by this rule to avoid duplicate lints.
-// It should be reported as a new pub type that is part of the crate's API.
-// This might seem like a nonsensical test but is a valid edge case.
-
-#[deprecated]
-pub union DeprecatedNewUnion {
-    bar: u64,
-}
+pub mod enums;
+pub mod structs;
+pub mod unions;
