+++ test_outputs/type_marked_deprecated.output.ron
@@ -1,39 +1,16 @@
 {
     "./test_crates/type_marked_deprecated/": [
         {
-            "name": String("StructToDeprecatedStruct"),
-            "new_attr": String("#[deprecated]"),
-            "owner_type": String("Struct"),
-            "path": List([
-                String("type_marked_deprecated"),
-                String("StructToDeprecatedStruct"),
-            ]),
-            "span_begin_line": Uint64(7),
-            "span_filename": String("src/lib.rs"),
-            "visibility_limit": String("public"),
-        },
-        {
-            "name": String("StructToDeprecatedMessageStruct"),
-            "new_attr": String("#[deprecated = \"This attribute was added\"]"),
-            "owner_type": String("Struct"),
-            "path": List([
-                String("type_marked_deprecated"),
-                String("StructToDeprecatedMessageStruct"),
-            ]),
-            "span_begin_line": Uint64(12),
-            "span_filename": String("src/lib.rs"),
-            "visibility_limit": String("public"),
-        },
-        {
             "name": String("EnumToDeprecatedEnum"),
             "new_attr": String("#[deprecated]"),
             "owner_type": String("Enum"),
             "path": List([
                 String("type_marked_deprecated"),
+                String("enums"),
                 String("EnumToDeprecatedEnum"),
             ]),
-            "span_begin_line": Uint64(69),
-            "span_filename": String("src/lib.rs"),
+            "span_begin_line": Uint64(5),
+            "span_filename": String("src/enums.rs"),
             "visibility_limit": String("public"),
         },
         {
@@ -42,10 +19,37 @@
             "owner_type": String("Enum"),
             "path": List([
                 String("type_marked_deprecated"),
+                String("enums"),
                 String("EnumToDeprecatedMessageEnum"),
             ]),
-            "span_begin_line": Uint64(74),
-            "span_filename": String("src/lib.rs"),
+            "span_begin_line": Uint64(10),
+            "span_filename": String("src/enums.rs"),
+            "visibility_limit": String("public"),
+        },
+        {
+            "name": String("StructToDeprecatedStruct"),
+            "new_attr": String("#[deprecated]"),
+            "owner_type": String("Struct"),
+            "path": List([
+                String("type_marked_deprecated"),
+                String("structs"),
+                String("StructToDeprecatedStruct"),
+            ]),
+            "span_begin_line": Uint64(5),
+            "span_filename": String("src/structs.rs"),
+            "visibility_limit": String("public"),
+        },
+        {
+            "name": String("StructToDeprecatedMessageStruct"),
+            "new_attr": String("#[deprecated = \"This attribute was added\"]"),
+            "owner_type": String("Struct"),
+            "path": List([
+                String("type_marked_deprecated"),
+                String("structs"),
+                String("StructToDeprecatedMessageStruct"),
+            ]),
+            "span_begin_line": Uint64(10),
+            "span_filename": String("src/structs.rs"),
             "visibility_limit": String("public"),
         },
     ]
